pass Int instead of Double in "Set Idle Limit" request

+++ transport/src/main/kotlin/net/yupol/transmissionremote/transport/rpc/RpcArgs.kt
@@ -161,6 +161,6 @@ object TorrentParameters {
             Parameter("seedIdleMode", mode.value)
 
     @JvmStatic
-    fun seedIdleLimit(limit: Double) =
+    fun seedIdleLimit(limit: Int) =
             Parameter("seedIdleLimit", limit)
 }
